+++ src/NetAppFiles/NetAppFiles/BackupPolicy/NewNetAppFilesBackupPolicy.cs
@@ -25,6 +25,7 @@ using Microsoft.Azure.Management.Monitor.Version2018_09_01.Models;
 using System.Collections.Generic;
 using System;
 using Microsoft.Azure.Commands.Common.Exceptions;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
 namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
 {
@@ -33,6 +34,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "NetAppFilesBackupPolicy",
         SupportsShouldProcess = true,
         DefaultParameterSetName = FieldsParameterSet), OutputType(typeof(PSNetAppFilesBackupPolicy))]
+    [CmdletOutputBreakingChange(typeof(PSNetAppFilesBackupPolicy), DeprecatedOutputProperties = new string[] { "YearlyBackupsToKeep" })]
     [Alias("New-AnfBackupPolicy")]
     public class NewAzureRmNetAppFilesBackupPolicy : AzureNetAppFilesCmdletBase
     {
@@ -100,6 +102,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
         [Parameter(
             Mandatory = false,
             HelpMessage = "Yearly backups count to keep")]
+        [CmdletParameterBreakingChange("YearlyBackupsToKeep", ChangeDescription = "Parameter YearlyBackupsToKeep is invalid and preserved for compatibility.")]
         [ValidateNotNullOrEmpty]
         public int? YearlyBackupsToKeep { get; set; }
 
@@ -161,8 +164,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.BackupPolicy
                 Tags = tagPairs,
                 DailyBackupsToKeep = DailyBackupsToKeep,
                 WeeklyBackupsToKeep = WeeklyBackupsToKeep,
-                MonthlyBackupsToKeep = MonthlyBackupsToKeep,
-                YearlyBackupsToKeep = YearlyBackupsToKeep
+                MonthlyBackupsToKeep = MonthlyBackupsToKeep
             };
 
             if (ShouldProcess(Name, string.Format(PowerShell.Cmdlets.NetAppFiles.Properties.Resources.CreateResourceMessage, ResourceGroupName)))