+++ src/NetAppFiles/NetAppFiles/Volume/NewNetAppFilesVolume.cs
@@ -117,7 +117,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
             Mandatory = true,
             HelpMessage = "The service level of the ANF volume")]
         [ValidateNotNullOrEmpty]
-        [PSArgumentCompleter("Standard", "Premium", "Ultra")]
+        [PSArgumentCompleter("Standard", "Premium", "Ultra", "StandardZRS")]
         public string ServiceLevel { get; set; }
 
         [Parameter(
@@ -149,7 +149,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
         [Parameter(
             Mandatory = false,
             HelpMessage = "Snapshot Policy ResourceId used to apply a snapshot policy to the volume")]
-        [ValidateNotNullOrEmpty]       
+        [ValidateNotNullOrEmpty]
         public string SnapshotPolicyId { get; set; }
         
         [Parameter(
@@ -224,6 +224,40 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Specifies whether the volume is enabled for Azure VMware Solution (AVS) datastore purpose (Enabled, Disabled)")]
+        [PSArgumentCompleter("Enabled", "Disabled")]
+        public string AvsDataStore { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Specifies if default quota is enabled for the volume")]
+        public SwitchParameter IsDefaultQuotaEnabled { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Default user quota for volume in KiBs. If isDefaultQuotaEnabled is set, the minimum value of 4 KiBs applies.")]
+        public long? DefaultUserQuotaInKiB { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Default group quota for volume in KiBs. If isDefaultQuotaEnabled is set, the minimum value of 4 KiBs applies.")]
+        public long? DefaultGroupQuotaInKiB { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Basic network, or Standard features available to the volume (Basic, Standard).")]
+        [PSArgumentCompleter("Basic", "Standard")]
+        public string NetworkFeature { get; set; }
+
+        [Parameter(
+            ParameterSetName = FieldsParameterSet,
+            Mandatory = false,
+            HelpMessage = "Network Sibling Set ID (GUID) for the group of volumes sharing networking resources example (9760acf5-4638-11e7-9bdb-020073ca3333).")]
+        [ValidateNotNullOrEmpty]
+        public string NetworkSiblingSetId { get; set; }
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "A hashtable which represents resource tags")]
         [ValidateNotNullOrEmpty]
         [Alias("Tags")]
@@ -268,7 +302,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
                     Replication = ReplicationObject,
                     Snapshot = new PSNetAppFilesVolumeSnapshot() { SnapshotPolicyId = SnapshotPolicyId },
                     Backup = Backup
-                };                
+                };
             }
 
             var volumeBody = new Management.NetApp.Models.Volume()
@@ -294,7 +328,12 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Volume
                 LdapEnabled = LdapEnabled,
                 CoolAccess = CoolAccess,
                 CoolnessPeriod = CoolnessPeriod,
-                UnixPermissions = UnixPermissions
+                UnixPermissions = UnixPermissions,
+                AvsDataStore = AvsDataStore,
+                IsDefaultQuotaEnabled = IsDefaultQuotaEnabled,
+                DefaultUserQuotaInKiBs = DefaultUserQuotaInKiB,
+                DefaultGroupQuotaInKiBs = DefaultGroupQuotaInKiB,
+                NetworkFeatures = NetworkFeature
             };
 
             if (ShouldProcess(Name, string.Format(PowerShell.Cmdlets.NetAppFiles.Properties.Resources.CreateResourceMessage, ResourceGroupName)))