+++ src/Resources/Resources/ChangeLog.md
@@ -19,12 +19,14 @@
 -->
 
 ## Upcoming Release
+* Added property `UIFormDefinition` to Template Spec Versions
+* `Export-AzTemplateSpec` will now include a Template Spec Version's UIFormDefinition (if any) as part of the export.
+
+## Version 4.4.0
 * Added a clearer error message for a case in which TemplateUri do not accept bicep file.
 * Fixed typos with ManagementGroups breaking change descriptions [#15819].
 * Fixed resource tags casing issue - resource tags casing not being preserved.
-* Update to Microsoft.Azure.Management.Authorization 2.13.0-preview.
-* Added property `UIFormDefinition` to Template Spec Versions
-* `Export-AzTemplateSpec` will now include a Template Spec Version's UIFormDefinition (if any) as part of the export.
+* Updated to Microsoft.Azure.Management.Authorization 2.13.0-preview.
 
 ## Version 4.3.1
 * Use JsonExtensions to serialize deserialize JSON objects to ensure the use of custom serialization settings [#15552]